Update debugging
Taking a closer look at the error message, GitPython is attempting to clone https://git@github.com:samhinshaw/dsci-100.git.git, an invalid url. It seems to be taking my SSH url and tacking on https:// and .git. I should check the options to see how to specify that we are using an SSH url