Python really likes for loops
If this were in JS, I would write one function which would map over the array and eventually reduce to the relevant element. But lambda functions in python are really limited (unlike JS’ arrow functions), and make this approach more complex. I suppose a reduce would still work here, but I found a pythonic method with for loops which gets the job done.